requirements.txt
streamlit
faiss-cpu
python-docx
numpy
PyPDF2
langchain
langchain-community
langchain-huggingface
sentence-transformers
huggingface_hub
